content/SEO для инфобизнеса актуальность, возможности, инструменты/SEO_РАБОТА/requirements.txt
# БАЗОВЫЕ ЗАВИСИМОСТИ ДЛЯ PYTHON + SELENIUM ПАРСЕРА
# Проект: Автоматизация анализа конкурентов для SEO
# Дата: 22.01.2026
# Python: 3.9+

# === SELENIUM ECOSYSTEM ===
selenium==4.17.0                # Браузерная автоматизация
undetected-chromedriver==3.5.5  # Обход антибот-защиты Яндекса (КРИТИЧНО!)
selenium-stealth==1.0.6         # Скрытие следов автоматизации
webdriver-manager==4.0.1        # Автообновление драйверов Chrome

# === ПАРСИНГ HTML ===
beautifulsoup4==4.12.3          # HTML парсинг
lxml==5.1.0                     # Быстрый парсер для BeautifulSoup
requests==2.31.0                # HTTP запросы для статических страниц

# === ОБРАБОТКА ДАННЫХ ===
pandas==2.2.0                   # Работа с таблицами и DataFrame
openpyxl==3.1.2                 # Экспорт в Excel (.xlsx)

# === УТИЛИТЫ ===
python-dotenv==1.0.1            # Управление переменными окружения (.env)
fake-useragent==1.5.1           # Генерация реалистичных User-Agent

# === ОПЦИОНАЛЬНО (для будущих фаз) ===
# anthropic==0.18.0             # Claude API для AI-анализа (Фаза 2)
# 2captcha-python==1.2.0        # Решение CAPTCHA (если понадобится)
